+++ src/Api/ResponseApi.php
@@ -17,6 +17,7 @@ namespace Katana\Sdk\Api;
 
 use Katana\Sdk\Api\Protocol\Http\HttpRequest;
 use Katana\Sdk\Api\Protocol\Http\HttpResponse;
+use Katana\Sdk\Api\Value\PayloadMeta;
 use Katana\Sdk\Api\Value\ReturnValue;
 use Katana\Sdk\Component\Component;
 use Katana\Sdk\Exception\InvalidValueException;
@@ -43,14 +44,9 @@ class ResponseApi extends Api implements Response
     private $transport;
 
     /**
-     * @var string
+     * @var PayloadMeta
      */
-    private $protocol;
-
-    /**
-     * @var string
-     */
-    private $gatewayAddress;
+    private $payloadMeta;
 
     /**
      * @var ReturnValue
@@ -71,8 +67,7 @@ class ResponseApi extends Api implements Response
      * @param HttpRequest $request
      * @param HttpResponse $response
      * @param Transport $transport
-     * @param string $protocol
-     * @param string $gatewayAddress,
+     * @param PayloadMeta $payloadMeta
      * @param ReturnValue $return
      */
     public function __construct(
@@ -88,8 +83,7 @@ class ResponseApi extends Api implements Response
         HttpRequest $request,
         HttpResponse $response,
         Transport $transport,
-        $protocol,
-        $gatewayAddress,
+        PayloadMeta $payloadMeta,
         ReturnValue $return
     ) {
         parent::__construct(
@@ -106,8 +100,7 @@ class ResponseApi extends Api implements Response
         $this->request = $request;
         $this->response = $response;
         $this->transport = $transport;
-        $this->protocol = $protocol;
-        $this->gatewayAddress = $gatewayAddress;
+        $this->payloadMeta = $payloadMeta;
         $this->return = $return;
     }
 
@@ -140,7 +133,7 @@ class ResponseApi extends Api implements Response
      */
     public function getGatewayProtocol(): string
     {
-        return $this->protocol;
+        return $this->payloadMeta->getGatewayProtocol();
     }
 
     /**
@@ -148,7 +141,7 @@ class ResponseApi extends Api implements Response
      */
     public function getGatewayAddress(): string
     {
-        return $this->gatewayAddress;
+        return $this->payloadMeta->getGatewayAddress();
     }
 
     /**
@@ -176,5 +169,23 @@ class ResponseApi extends Api implements Response
                 $action
             ));
         }
+    }
+
+    /**
+     * @param string $name
+     * @param string $default
+     * @return string
+     */
+    public function getRequestAttribute(string $name, string $default = ''): string
+    {
+        return $this->payloadMeta->getAttribute($name, $default);
+    }
+
+    /**
+     * @return array
+     */
+    public function getRequestAttributes(): array
+    {
+        return $this->payloadMeta->getAttributes();
     }
 }